backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "shahdhesham5", email = "shahd.hesham1605@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.6",
    "django-allauth>=65.11.2",
    "django-cors-headers>=4.8.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "face-recognition>=1.3.0",
    "face-recognition-models",
    "opencv-python>=4.12.0.88",
    "openpyxl>=3.1.5",
    "psycopg[binary]>=3.2.10",
    "python-bidi>=0.5.0",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models.git" }
